new platform mqtt_json
